requirements.txt
fastapi==0.115.5
numpy==1.26.4
music21==9.3.0
mysql-connector-python==9.1.0
pretty_midi==0.2.10
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
uvicorn==0.27.1
scikit-learn==1.2.2
keras==3.7.0
pillow==10.2.0
python-multipart==0.0.17
<<<<<<< HEAD
pydantic==1.10.12
=======
pydantic==1.10.12
>>>>>>> 757afcd (update DB and requirement)
